feat: circles-plus-plus
* fix: correct movement calculation

Note: Implemented the basic wanted functionality, but more needs to be
done, such as making sure I'm not abusing memory, cleaning up the code,
and creating classes with better structuring. Also need to find a way to
have the executable to open without cli and test on different computers,
and eventually run this on linux.